feat: move files to use as a module, fix relative imports
