+++ pff/pitch_control/cpitchcontrol.pyx
